Handle computed key meta
This seems overly complicated. Probably it's best to just always use the
hash-based key method, for simplicity, but it's certainly harder to
debug. Not sure what the best way forward is right now, so I'll just
keep everything for now.

Worst case, if we remove the :key support in the future it shouldn't
break existing clients since it's just meta data.