Add support for top-level style composition
Refs #1

Uses the same `composes` key that css-modules uses.

This is actually pretty straightforward. By just using the method
call within the style factory, we know that the style element will
be inserted, and we are given the class name directly, so we can
just insert that into the string of classes returned from our
method.

It'd be cool to support :composes in child definitions, but since
it's implemented by concat'ing class names it doesn't seem possible.
For that purpose you would just create a separate def* and use it
on the children directly.