Make defkeyframes accept a params vector
This is a breaking change! But it makes the API much more logically
consistent—why shouldn't you be able to parameterize keyframes?

This change also includes some heavy refactoring to simplify code
generation for the other cases, as well as an optimization that skips
the call to build-style-name if the function does not accept parameters.