Don't touch the DOM unless a style has actually changed
I noticed that when using a spade defattr on an item within a
virtualized list, scroll performance got a bit janky. Other virtualized
list components performed just fine, so I dove into the profiler to see
what was going on.

It turns out that each call to the defattr function was touching the
dom, even though its params weren't changed. The reason I wasn't running
into this with other virtualized lists was that they were styling their
children at a higher level, so the function was only called once.

So, this change will only touch the dom if the underlying style has
actually changed. Since Spade encourages small styles, this comparison
should be fast enough---and certainly faster than updating the DOM---but
if anyone discovers performance issues we might be able to assume that
in production, any given style function, when provided the same
parameters, will always produce the same CSS, and thus skip updates
entirely in production, while keeping the current behavior for testing
styles with hot reload in dev mode.