Add initial tests for and experiments with multi-return binding forms
See #13

This is problematic when used in nested forms. We could perhaps add a
special case so only top-level forms are supported, but that might be
more confusing than it's worth. Another alternative could be some
special syntax that triggers us to unpack binding forms (IE: explicitly
opt-in to multi-return) but that might be a bit too arcane, eg:

```clojure
(defclass stylish []
  [:* (let [opacity 0.2] ... )])
```

We could also add special binding form replacements to similarly opt-in,
eg:

```clojure
(defclass stylish []
  ; the & mimics [:&]
  (&let [opacity 0.2] ... ))
```

This is all sugar though, so maybe the real solution is just
documentation?